requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
typing_extensions==4.11.0
djangorestframework~=3.15.1
django-filter~=24.2
requests~=2.31.0
celery~=5.4.0
